Move play function to hook
